+++ MidtermProject/denseGEMM-SUMMA/src/utils.c
@@ -4,22 +4,22 @@
 #include <stdlib.h>
 #include <string.h>
 
-void matmul(float *A, float *B, float *C, int m, int n, int k)
+void matmul(double *A, double *B, double *C, int m, int n, int k)
 {
     // Initialize output matrix to zero
-    memset(C, 0, m * n * sizeof(float));
+    memset(C, 0, m * n * sizeof(double));
 
     // C[i,j] = sum(A[i,p] * B[p,j])
     for (int i = 0; i < m; i++)
     {
         for (int j = 0; j < n; j++)
         {
-            float sum = 0.0f;
+            double sum = 0.0f;
             for (int p = 0; p < k; p++)
             {
                 // A is m×k, B is k×n
-                float a_val = A[i * k + p]; // A[i,p]
-                float b_val = B[p * n + j]; // B[p,j]
+                double a_val = A[i * k + p]; // A[i,p]
+                double b_val = B[p * n + j]; // B[p,j]
                 sum += a_val * b_val;
             }
             C[i * n + j] = sum;
@@ -27,12 +27,12 @@ void matmul(float *A, float *B, float *C, int m, int n, int k)
     }
 }
 
-void verify_result(float *C_global, float *A, float *B, int m, int n, int k)
+void verify_result(double *C_global, double *A, double *B, int m, int n, int k)
 {
     int errors = 0;
-    float tolerance = 1e-5;
+    double tolerance = 1e-5;
     // Perform reference matrix multiplication
-    float *C_ref = (float *)calloc(m * n, sizeof(float));
+    double *C_ref = (double *)calloc(m * n, sizeof(double));
     if (!C_ref)
     {
         printf("Error: Failed to allocate memory for C_ref\n");
@@ -44,7 +44,7 @@ void verify_result(float *C_global, float *A, float *B, int m, int n, int k)
     {
         for (int j = 0; j < n; j++)
         {
-            float sum = 0.0f;
+            double sum = 0.0f;
             for (int p = 0; p < k; p++)
             {
                 sum += A[i * k + p] * B[p * n + j];
@@ -54,8 +54,8 @@ void verify_result(float *C_global, float *A, float *B, int m, int n, int k)
     }
 
     // Compute detailed error statistics
-    float max_error = 0.0f;
-    float avg_error = 0.0f;
+    double max_error = 0.0f;
+    double avg_error = 0.0f;
     int max_error_index = -1;
     int max_error_i = -1, max_error_j = -1;
 
@@ -64,7 +64,7 @@ void verify_result(float *C_global, float *A, float *B, int m, int n, int k)
         for (int j = 0; j < n; j++)
         {
             int idx = i * n + j;
-            float curr_error = fabs(C_global[idx] - C_ref[idx]);
+            double curr_error = fabs(C_global[idx] - C_ref[idx]);
             avg_error += curr_error;
 
             if (curr_error > max_error)
